refactor!: Rename states -> projects
