+++ docs/projects/git.md
@@ -1,22 +1,22 @@
-# `libvcs.states.git`
+# `libvcs.projects.git`
 
 For `git(1)`.
 
 Compare to:
 [`fabtools.require.git`](https://fabtools.readthedocs.io/en/0.19.0/api/require/git.html),
-[`salt.states.git`](https://docs.saltproject.io/en/latest/ref/states/all/salt.states.git.html),
+[`salt.projects.git`](https://docs.saltproject.io/en/latest/ref/projects/all/salt.projects.git.html),
 [`ansible.builtin.git`](https://docs.ansible.com/ansible/latest/collections/ansible/builtin/git_module.html)
 
 ```{eval-rst}
 .. autosummary::
    :recursive:
 
-   libvcs.states.git.GitRemote
-   libvcs.states.git.GitRepo
+   libvcs.projects.git.GitRemote
+   libvcs.projects.git.GitRepo
 ```
 
 ```{eval-rst}
-.. automodule:: libvcs.states.git
+.. automodule:: libvcs.projects.git
    :members:
    :show-inheritance:
    :undoc-members:
