+++ docs/projects/index.md
@@ -1,10 +1,10 @@
-(states)=
+(projects)=
 
-# `libvcs.states`
+# `libvcs.projects`
 
 Compare to:
 [`fabtools.require.git`](https://fabtools.readthedocs.io/en/0.19.0/api/require/git.html),
-[`salt.states.git`](https://docs.saltproject.io/en/latest/ref/states/all/salt.states.git.html),
+[`salt.projects.git`](https://docs.saltproject.io/en/latest/ref/projects/all/salt.projects.git.html),
 [`ansible.builtin.git`](https://docs.ansible.com/ansible/latest/collections/ansible/builtin/git_module.html)
 
 :::{warning}
@@ -38,7 +38,7 @@ See examples below of git, mercurial, and subversion.
 ## Constants
 
 ```{eval-rst}
-.. automodule:: libvcs.states.constants
+.. automodule:: libvcs.projects.constants
    :members:
 ```
