+++ docs/redirects.txt
@@ -1,5 +1,5 @@
-"api.md" "states/index.md"
-"api/base.md" "states/base.md"
-"api/git.md" "states/git.md"
-"api/hg.md" "states/hg.md"
-"api/svn.md" "states/svn.md"
+"api.md" "projects/index.md"
+"api/base.md" "projects/base.md"
+"api/git.md" "projects/git.md"
+"api/hg.md" "projects/hg.md"
+"api/svn.md" "projects/svn.md"
