+++ libvcs/projects/base.py
