+++ libvcs/projects/constants.py
