+++ libvcs/projects/git.py
@@ -21,7 +21,11 @@ from typing import Dict, NamedTuple, Optional, TypedDict, Union
 from urllib import parse as urlparse
 
 from .. import exc
-from ..states.base import BaseRepo, VCSLocation, convert_pip_url as base_convert_pip_url
+from ..projects.base import (
+    BaseRepo,
+    VCSLocation,
+    convert_pip_url as base_convert_pip_url,
+)
 
 logger = logging.getLogger(__name__)
