+++ tests/projects/test_base.py
@@ -4,8 +4,8 @@ import sys
 
 import pytest
 
+from libvcs.projects.base import BaseRepo, convert_pip_url
 from libvcs.shortcuts import create_repo
-from libvcs.states.base import BaseRepo, convert_pip_url
 
 
 def test_repr():
