+++ tests/projects/test_git.py
@@ -12,14 +12,14 @@ from pytest_mock import MockerFixture
 from libvcs import exc
 from libvcs.cmd.core import run, which
 from libvcs.conftest import CreateRepoCallbackFixtureProtocol
-from libvcs.shortcuts import create_repo_from_pip_url
-from libvcs.states.git import (
+from libvcs.projects.git import (
     GitFullRemoteDict,
     GitRemote,
     GitRepo,
     convert_pip_url as git_convert_pip_url,
     extract_status,
 )
+from libvcs.shortcuts import create_repo_from_pip_url
 
 if not which("git"):
     pytestmark = pytest.mark.skip(reason="git is not available")
