+++ tests/projects/test_svn.py
@@ -6,8 +6,8 @@ import pytest
 
 from libvcs.cmd.core import which
 from libvcs.conftest import CreateRepoCallbackFixtureProtocol
+from libvcs.projects.svn import SubversionRepo
 from libvcs.shortcuts import create_repo_from_pip_url
-from libvcs.states.svn import SubversionRepo
 
 if not which("svn"):
     pytestmark = pytest.mark.skip(reason="svn is not available")
